move link related utils to a separate file
